refactor: 모바일 레이아웃 컴포넌트의 이름 변경
