+++ src/components/layout/MobileSideMenu/css/SideBarContent.module.css
