+++ src/components/layout/MobileSideMenu/index.tsx
@@ -1,7 +1,7 @@
 import { service } from '@/service/pickService';
 import SideBarContent from './SideBarContent';
 
-export default async function SideBar() {
+export default async function MobileSideMenu() {
   const categories = await service.category.getAllMainCategory();
 
   return <SideBarContent categories={categories} />;
