+++ src/components/layout/MobileTimerBar/Timer.tsx
