--- src/components/layout/SearchBar/SearchBar.tsx
@@ -1,8 +1,0 @@
-import { service } from '@/service/pickService';
-import SearchBarcontent from './SearchBarcontent';
-
-export default async function SearchBar() {
-  const monsters = await service.search.getMonsters();
-
-  return <SearchBarcontent monsters={monsters} />;
-}
